Move agent-manager related code to kilocode directories
