+++ src/core/kilocode/agent-manager/AgentRegistry.ts
