+++ src/core/kilocode/agent-manager/CliOutputParser.ts
