+++ src/core/kilocode/agent-manager/__tests__/AgentManagerProvider.spec.ts
