+++ src/core/kilocode/agent-manager/__tests__/CliOutputParser.test.ts
