+++ src/core/kilocode/agent-manager/types.ts
