+++ webview-ui/agent-manager.html
@@ -7,6 +7,6 @@
 	</head>
 	<body>
 		<div id="root"></div>
-		<script type="module" src="/src/agent-manager/index.tsx"></script>
+		<script type="module" src="/src/kilocode/agent-manager/index.tsx"></script>
 	</body>
 </html>
