+++ webview-ui/src/kilocode/agent-manager/components/AgentManagerApp.css
@@ -120,7 +120,6 @@
 	text-align: center;
 }
 
-
 .status-icon.done {
 	color: var(--vscode-charts-green);
 }
