+++ webview-ui/src/kilocode/agent-manager/components/AgentManagerApp.tsx
