+++ webview-ui/src/kilocode/agent-manager/components/ChatInput.tsx
