+++ webview-ui/src/kilocode/agent-manager/components/SessionSidebar.tsx
