+++ webview-ui/src/kilocode/agent-manager/components/index.ts
