+++ webview-ui/src/kilocode/agent-manager/state/__tests__/multi-agent.spec.ts
