+++ webview-ui/src/kilocode/agent-manager/state/atoms/index.ts
