+++ webview-ui/src/kilocode/agent-manager/state/atoms/sessions.ts
