+++ webview-ui/src/kilocode/agent-manager/state/hooks/index.ts
