+++ webview-ui/src/kilocode/agent-manager/state/index.ts
