+++ webview-ui/src/kilocode/agent-manager/utils/vscode.ts
